iss #507: rename smallboard => cardboard
