+++ Procfile
@@ -1,4 +1,4 @@
 release: python manage.py migrate --noinput
-web: gunicorn smallboard.wsgi --log-file -
-worker: celery -A smallboard worker -l INFO --without-heartbeat --without-gossip --without-mingle
+web: gunicorn cardboard.wsgi --log-file -
+worker: celery -A cardboard worker -l INFO --without-heartbeat --without-gossip --without-mingle
 bot: python manage.py rundiscordbot
