+++ cardboard/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for smallboard project.
+Django settings for cardboard project.
 
 Generated by 'django-admin startproject' using Django 2.2.7.
 
@@ -82,12 +82,12 @@ MIDDLEWARE = [
     "social_django.middleware.SocialAuthExceptionMiddleware",
 ]
 
-ROOT_URLCONF = "smallboard.urls"
+ROOT_URLCONF = "cardboard.urls"
 
 TEMPLATES = [
     {
         "BACKEND": "django.template.backends.django.DjangoTemplates",
-        "DIRS": ["smallboard/templates/"],
+        "DIRS": ["cardboard/templates/"],
         "APP_DIRS": True,
         "OPTIONS": {
             "context_processors": [
@@ -97,13 +97,13 @@ TEMPLATES = [
                 "django.contrib.messages.context_processors.messages",
                 "social_django.context_processors.backends",
                 "social_django.context_processors.login_redirect",
-                "smallboard.context_processors.google_auth",
+                "cardboard.context_processors.google_auth",
             ],
         },
     },
 ]
 
-WSGI_APPLICATION = "smallboard.wsgi.application"
+WSGI_APPLICATION = "cardboard.wsgi.application"
 
 # Password validation
 # https://docs.djangoproject.com/en/2.2/ref/settings/#auth-password-validators
@@ -143,7 +143,7 @@ USE_TZ = True
 
 STATIC_URL = "/static/"
 STATICFILES_DIRS = [
-    os.path.join(BASE_DIR, "smallboard/static"),
+    os.path.join(BASE_DIR, "cardboard/static"),
     os.path.join(BASE_DIR, "hunts/static"),
 ]
 STATICFILES_STORAGE = "whitenoise.django.CompressedManifestStaticFilesStorage"
@@ -171,7 +171,7 @@ import dj_database_url
 
 DATABASES = {}
 DATABASES["default"] = dj_database_url.config(conn_max_age=600, ssl_require=False)
-DATABASES["default"]["TEST"] = {"NAME": "test_smallboard"}
+DATABASES["default"]["TEST"] = {"NAME": "test_cardboard"}
 
 # Discord API. See
 # https://support.discord.com/hc/en-us/articles/206346498-Where-can-I-find-my-User-Server-Message-ID-
@@ -225,7 +225,7 @@ AUTHENTICATION_BACKENDS = [
 
 SOCIAL_AUTH_URL_NAMESPACE = "social"
 
-# smallboard client id
+# cardboard client id
 try:
     SOCIAL_AUTH_GOOGLE_OAUTH2_KEY = os.environ["SOCIAL_AUTH_GOOGLE_OAUTH2_KEY"]
     SOCIAL_AUTH_GOOGLE_OAUTH2_SECRET = os.environ["SOCIAL_AUTH_GOOGLE_OAUTH2_SECRET"]
