+++ cardboard/static/favicon.ico
