+++ cardboard/templates/base.html
@@ -1,7 +1,7 @@
 {% load static %}
 <html lang="en">
     <head>
-        <title>Cardinality Smallboard</title>
+        <title>Cardinality Cardboard</title>
         <link rel="shortcut icon" type="image/png" href="{% static 'favicon.ico' %}"/>
 
         <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
@@ -16,7 +16,7 @@
         <nav class="navbar navbar-expand-lg navbar-light bg-light">
             <a class="navbar-brand" href="/">
                 <img height="40" src="{% static 'favicon.ico' %}" alt="">
-                Smallboard
+                Cardboard
             </a>
             <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
             <span class="navbar-toggler-icon"></span>
