+++ cardboard/templates/home.html
@@ -8,7 +8,7 @@
     <div class="row">
         <div class="col-sm">
 
-            <h1>Cardinality Smallboard</h1>
+            <h1>Cardinality Cardboard</h1>
 
 
             <ul>
@@ -24,7 +24,7 @@
                 </li>
                 <li>
                     <b>How are you using my data?</b><br>
-                    We match your email against a whitelist of emails for each hunt. The actions you take while using the smallboard (such as solving a puzzle) may be logged and attributed to your email. This data is only used for logging purposes. See <a href="{% url 'privacy'%}">here</a> for our full privacy policy.
+                    We match your email against a whitelist of emails for each hunt. The actions you take while using the cardboard (such as solving a puzzle) may be logged and attributed to your email. This data is only used for logging purposes. See <a href="{% url 'privacy'%}">here</a> for our full privacy policy.
                 </li>
 
                 <li>
