+++ cardboard/templates/privacy.html
