+++ cardboard/urls.py
@@ -1,4 +1,4 @@
-"""smallboard URL Configuration
+"""cardboard URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/2.2/topics/http/urls/
