+++ cardboard/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for smallboard project.
+WSGI config for cardboard project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "smallboard.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "cardboard.settings")
 
 application = get_wsgi_application()
