+++ docker-compose.yaml
@@ -6,21 +6,21 @@ services:
     volumes:
       - postgres_data:/var/lib/postgresql/data
     environment:
-      - POSTGRES_USER=smallboard_user
-      - POSTGRES_PASSWORD=smallboard_password
-      - POSTGRES_DB=smallboard
+      - POSTGRES_USER=cardboard_user
+      - POSTGRES_PASSWORD=cardboard_password
+      - POSTGRES_DB=cardboard
   redis:
     image: redis:alpine
   web:
     build: .
     command: python manage.py runserver 0.0.0.0:8000
     volumes:
-      - .:/usr/src/smallboard/:Z
-      - /usr/src/smallboard/node_modules
+      - .:/usr/src/cardboard/:Z
+      - /usr/src/cardboard/node_modules
     ports:
       - 8000:8000
     environment:
-      - DATABASE_URL=postgres://smallboard_user:smallboard_password@db/smallboard
+      - DATABASE_URL=postgres://cardboard_user:cardboard_password@db/cardboard
       - REDIS_URL=redis://redis:6379
     env_file:
       - ./.env
@@ -31,11 +31,11 @@ services:
       - redis
   celery:
     build: .
-    command: celery -A smallboard worker -l info
+    command: celery -A cardboard worker -l info
     volumes:
-      - .:/usr/src/smallboard/:Z
+      - .:/usr/src/cardboard/:Z
     environment:
-      - DATABASE_URL=postgres://smallboard_user:smallboard_password@db/smallboard
+      - DATABASE_URL=postgres://cardboard_user:cardboard_password@db/cardboard
       - REDIS_URL=redis://redis:6379
     env_file:
       - ./.env
